[8.0] Fix various deprecation warnings

+++ src/DIRAC/StorageManagementSystem/DB/StorageManagementDB.py
@@ -9,7 +9,6 @@
     The TaskReplicas table maps the TaskIDs from the Tasks table to the ReplicaID from the CacheReplicas table.
     The StageRequests table contains each of the prestage request IDs for each of the replicas.
 """
-from past.builtins import long
 import six
 import inspect
 import threading
@@ -1337,7 +1336,7 @@ class StorageManagementDB(DB):
         resSummary = {}
         i = 1
         for status, se, numFiles, sumFiles in res["Value"]:
-            resSummary[i] = {"Status": status, "SE": se, "NumFiles": long(numFiles), "SumFiles": float(sumFiles)}
+            resSummary[i] = {"Status": status, "SE": se, "NumFiles": int(numFiles), "SumFiles": float(sumFiles)}
             i += 1
         return S_OK(resSummary)
